fix: account for decimals, return double

+++ app/model/LargestNumberResult.scala
@@ -2,7 +2,7 @@ package model
 
 import play.api.libs.json.{Json, OFormat}
 
-case class LargestNumberResult(fileName: String, largestNumber: Option[Long])
+case class LargestNumberResult(fileName: String, largestNumber: Option[Double])
 object LargestNumberResult {
   implicit val format: OFormat[LargestNumberResult] = Json.format[LargestNumberResult]
 }
